+++ lib/grinder.dart
@@ -40,7 +40,7 @@ export 'src/tools.dart';
 /// run from a project root.
 ///
 /// If a task fails, it throws and runs no further tasks.
-Future grind(List<String> args, {bool verifyProjectRoot = true}) {
+Future<void> grind(List<String> args, {bool verifyProjectRoot = true}) {
   try {
     discoverTasks(grinder, currentMirrorSystem().isolate.rootLibrary);
     return runTasks(args, verifyProjectRoot: verifyProjectRoot);
@@ -51,24 +51,4 @@ Future grind(List<String> args, {bool verifyProjectRoot = true}) {
 
     return Future.error(e);
   }
-}
-
-/// Start the build process. This should be called at the end of the `main()`
-/// method. If there is a task failure, this method will halt task execution and
-/// throw.
-@Deprecated('Use `grind` instead.')
-Future startGrinder(List<String> args, {bool verifyProjectRoot = true}) {
-  return runTasks(args, verifyProjectRoot: verifyProjectRoot);
-}
-
-/// Add a new task definition to the global [Grinder] instance. A [name] is
-/// required. If specified, a [taskFunction] is invoked when the task starts.
-/// Any dependencies of the task, that need to run before it, should be passed
-/// in via [depends].
-@Deprecated('Use the task annotations instead.')
-void task(String name,
-    [Function? taskFunction, List<String> depends = const []]) {
-  grinder.addTask(GrinderTask(name,
-      taskFunction: taskFunction,
-      depends: depends.map((dep) => TaskInvocation(dep))));
 }